MVC folder restructure and working sign in/sign up system
