+++ src/app/models/user/user.go
@@ -1,192 +1,1 @@
-package user
-
-import(
-	"log"
-	r "github.com/dancannon/gorethink"
-	"encoding/json"
-	"strconv"
-)
-
-var session *r.Session
-
-type User struct{
-	Id string `gorethink:"id,omitempty"`
-	Username string
-	Email	string
-	Password string
-}
-
-func Init() {
-    var err error
-    session, err = r.Connect(r.ConnectOpts{
-        Address:  "localhost:28015",
-        Database: "ease",
-    })
-
-    if err != nil {
-        log.Println(err)
-        return
-    }
-}
-
-
-
-func FindUser(Id string) (string){
-	result, err := r.DB("ease").Table("users").Get(Id).Run(session)
-	if err != nil{
-		log.Println(err)
-	}
-	return printObj(result)
-}
-
-func Insert() string{
-	var data = map[string]interface{}{
-        "Name":  "David Davidson",
-        "Place": "Somewhere",
-  }
-	result, err := r.DB("ease").Table("users").Insert(data).RunWrite(session)
-	if err != nil{
-		log.Println(err)
-		return ""
-	}
-	log.Println(printObj(result))
-	return result.GeneratedKeys[0]
-}
-
-
-func RecordCount(db string) string {
-    cursor, err := r.DB("ease").Table(db).Count().Run(session)
-    if err != nil {
-        log.Println(err)
-        return strconv.Itoa(-1)
-    }
-
-    var cnt int
-    cursor.One(&cnt)
-    cursor.Close()
-
-    return printObj(cnt)
-}
-
-func GetNth(n int) (string){
-	result, err := r.DB("ease").Table("users").Limit(n).Run(session)
-	if err != nil{
-		log.Println(err)
-	}
-	return printObj(result)
-}
-
-func printObj(v interface{}) (string) {
-    vBytes, _ := json.Marshal(v)
-    return (string(vBytes))
-}
-
-func FetchAll(table string) string{
-    rows, err := r.Table(table).Run(session)
-    if err != nil {
-        log.Println(err)
-        return "Table "+table+" doesn't exist"
-    }
-
-    // Read records into persons slice
-    var records []User
-    err2 := rows.All(&records)
-    if err2 != nil {
-        log.Println(err2)
-        return "error caught2"
-    }
-
-		result := ""
-    for _, p := range records {
-        result += printObj(p)
-    }
-		return result
-}
-
-/*
-
-import (
-	"crypto/rand"
-	"errors"
-	"log"
-	"time"
-
-	r "github.com/dancannon/gorethink"
-	"golang.org/x/crypto/bcrypt"
-)
-
-type User struct {
-	Id           string `gorethink:"id,omitempty"`
-	Username     string
-	PasswordHash string
-	ApiToken     string
-	CreatedAt    time.Time
-}
-
-func NewUser(username, password string) (*User, error) {
-	user := new(User)
-	user.CreatedAt = time.Now()
-	user.Username = username
-	randToken := make([]byte, 30)
-	_, err := rand.Read(randToken)
-	if err != nil {
-		log.Println("Error: Couldn't generate random API token.")
-		log.Println(err)
-		return nil, err
-	}
-	user.ApiToken = string(randToken)
-	byteHash, err :=
-		bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	user.PasswordHash = string(byteHash)
-	if err != nil {
-		log.Println("Error: Couldn't hash password.")
-		log.Println(err)
-		return nil, err
-	}
-	return user, nil
-}
-
-func (user *User) Save() error {
-	// Check that a user with the given username doesn't already exist.
-	otherUser := Find(user.Username)
-	if otherUser != nil && user.Id != otherUser.Id {
-		return errors.New("Error: A user with that name already exists.")
-	}
-
-	_, err := r.Table("users").Insert(user).RunWrite(db.Session)
-	if err != nil {
-		friendlyErr := errors.New("Error: Couldn't save user.")
-		log.Println(friendlyErr)
-		log.Println(err)
-		return friendlyErr
-	}
-	return nil
-}
-
-func Find(username string) *User {
-	res, err := r.Table("users").Filter(map[string]string{
-		"username": username,
-	}).Run(db.Session)
-	if err != nil || res.IsNil() {
-		return nil
-	}
-	var user *User
-	err = res.One(&user)
-	if err != nil {
-		return nil
-	}
-	return user
-}
-
-func AttemptLogin(username, password string) *User {
-	user := Find(username)
-	if user == nil {
-		return nil
-	}
-	err :=
-		bcrypt.CompareHashAndPassword([]byte(user.PasswordHash), []byte(password))
-	if err != nil {
-		return nil
-	}
-	return user
-}*/
+package models
